+++ apm-agent-core/src/main/java/co/elastic/apm/agent/report/serialize/DslJsonSerializer.java
@@ -52,7 +52,9 @@ import co.elastic.apm.agent.impl.transaction.Span;
 import co.elastic.apm.agent.impl.transaction.SpanCount;
 import co.elastic.apm.agent.impl.transaction.TraceContext;
 import co.elastic.apm.agent.impl.transaction.Transaction;
+import co.elastic.apm.agent.metrics.Labels;
 import co.elastic.apm.agent.metrics.MetricRegistry;
+import co.elastic.apm.agent.metrics.MetricSet;
 import co.elastic.apm.agent.util.PotentiallyMultiValuedMap;
 import com.dslplatform.json.BoolConverter;
 import com.dslplatform.json.DslJson;
@@ -81,7 +83,7 @@ import static com.dslplatform.json.JsonWriter.COMMA;
 import static com.dslplatform.json.JsonWriter.OBJECT_END;
 import static com.dslplatform.json.JsonWriter.OBJECT_START;
 
-public class DslJsonSerializer implements PayloadSerializer {
+public class DslJsonSerializer implements PayloadSerializer, MetricRegistry.MetricsReporter {
 
     /**
      * Matches default ZLIB buffer size.
@@ -229,8 +231,13 @@ public class DslJsonSerializer implements PayloadSerializer {
     }
 
     @Override
+    public void report(Map<? extends Labels, MetricSet> metricSets) {
+        MetricRegistrySerializer.serialize(metricSets, replaceBuilder, jw);
+    }
+
+    @Override
     public void serializeMetrics(MetricRegistry metricRegistry) {
-        MetricRegistrySerializer.serialize(metricRegistry, replaceBuilder, jw);
+        metricRegistry.report(this);
     }
 
     private void serializeErrorPayload(ErrorPayload payload) {
@@ -519,7 +526,7 @@ public class DslJsonSerializer implements PayloadSerializer {
         writeField("name", transaction.getName());
         serializeTraceContext(transaction.getTraceContext(), false);
         writeField("type", transaction.getType());
-        writeField("duration", transaction.getDuration());
+        writeField("duration", transaction.getDurationMs());
         writeField("result", transaction.getResult());
         serializeContext(transaction.getContext(), transaction.getTraceContext());
         serializeSpanCount(transaction.getSpanCount());
@@ -560,7 +567,7 @@ public class DslJsonSerializer implements PayloadSerializer {
         writeField("name", span.getName());
         writeTimestamp(span.getTimestamp());
         serializeTraceContext(span.getTraceContext(), true);
-        writeField("duration", span.getDuration());
+        writeField("duration", span.getDurationMs());
         if (span.getStacktrace() != null) {
             serializeStacktrace(span.getStacktrace().getStackTrace());
         }
@@ -809,6 +816,45 @@ public class DslJsonSerializer implements PayloadSerializer {
         jw.writeByte(OBJECT_END);
     }
 
+    static void serializeLabels(Labels labels, StringBuilder replaceBuilder, JsonWriter jw) {
+        if (!labels.isEmpty()) {
+            if (labels.getTransactionName() != null || labels.getTransactionType() != null) {
+                writeFieldName("transaction", jw);
+                jw.writeByte(OBJECT_START);
+                writeField("name", labels.getTransactionName(), replaceBuilder, jw);
+                writeLastField("type", labels.getTransactionType(), replaceBuilder, jw);
+                jw.writeByte(OBJECT_END);
+                jw.writeByte(COMMA);
+            }
+
+            if (labels.getSpanType() != null || labels.getSpanSubType() != null) {
+                writeFieldName("span", jw);
+                jw.writeByte(OBJECT_START);
+                writeField("type", labels.getSpanType(), replaceBuilder, jw);
+                writeLastField("subtype", labels.getSpanSubType(), replaceBuilder, jw);
+                jw.writeByte(OBJECT_END);
+                jw.writeByte(COMMA);
+            }
+
+            writeFieldName("tags", jw);
+            jw.writeByte(OBJECT_START);
+            serialize(labels, replaceBuilder, jw);
+            jw.writeByte(OBJECT_END);
+            jw.writeByte(COMMA);
+        }
+    }
+
+    private static void serialize(Labels labels, StringBuilder replaceBuilder, JsonWriter jw) {
+        for (int i = 0; i < labels.size(); i++) {
+            if (i > 0) {
+                jw.writeByte(COMMA);
+            }
+            writeStringValue(sanitizeLabelKey(labels.getKey(i), replaceBuilder), replaceBuilder, jw);
+            jw.writeByte(JsonWriter.SEMI);
+            serializeScalarValue(replaceBuilder, jw, labels.getValue(i), false);
+        }
+    }
+
     private static void serializeScalarValue(StringBuilder replaceBuilder, JsonWriter jw, Object value, boolean extendedStringLimit) {
         if (value instanceof String) {
             if (extendedStringLimit) {
@@ -983,9 +1029,13 @@ public class DslJsonSerializer implements PayloadSerializer {
     }
 
     void writeField(final String fieldName, @Nullable final String value) {
+        writeField(fieldName, value, replaceBuilder, jw);
+    }
+
+    static void writeField(final String fieldName, @Nullable final CharSequence value, StringBuilder replaceBuilder, JsonWriter jw) {
         if (value != null) {
-            writeFieldName(fieldName);
-            writeStringValue(value);
+            writeFieldName(fieldName, jw);
+            writeStringValue(value, replaceBuilder, jw);
             jw.writeByte(COMMA);
         }
     }
@@ -1073,9 +1123,13 @@ public class DslJsonSerializer implements PayloadSerializer {
     }
 
     void writeLastField(final String fieldName, @Nullable final String value) {
-        writeFieldName(fieldName);
+        writeLastField(fieldName, value, replaceBuilder, jw);
+    }
+
+    static void writeLastField(final String fieldName, @Nullable final String value, StringBuilder replaceBuilder, final JsonWriter jw) {
+        writeFieldName(fieldName, jw);
         if (value != null) {
-            writeStringValue(value);
+            writeStringValue(value, replaceBuilder, jw);
         } else {
             jw.writeNull();
         }