+++ apm-agent-core/src/test/java/co/elastic/apm/agent/TransactionUtils.java
@@ -24,7 +24,6 @@
  */
 package co.elastic.apm.agent;
 
-import co.elastic.apm.agent.impl.ElasticApmTracer;
 import co.elastic.apm.agent.impl.context.Request;
 import co.elastic.apm.agent.impl.context.TransactionContext;
 import co.elastic.apm.agent.impl.sampling.ConstantSampler;
@@ -34,8 +33,6 @@ import co.elastic.apm.agent.impl.transaction.Transaction;
 
 import java.util.ArrayList;
 import java.util.List;
-
-import static org.mockito.Mockito.mock;
 
 public class TransactionUtils {
 
@@ -86,7 +83,7 @@ public class TransactionUtils {
 
     public static List<Span> getSpans(Transaction t) {
         List<Span> spans = new ArrayList<>();
-        Span span = new Span(mock(ElasticApmTracer.class))
+        Span span = new Span(MockTracer.create())
             .start(TraceContext.fromParent(), t, -1, false)
             .withName("SELECT FROM product_types")
             .withType("db")
@@ -102,20 +99,20 @@ public class TransactionUtils {
         span.addLabel("framework", "some-framework");
         spans.add(span);
 
-        spans.add(new Span(mock(ElasticApmTracer.class))
+        spans.add(new Span(MockTracer.create())
             .start(TraceContext.fromParent(), t, -1, false)
             .withName("GET /api/types")
             .withType("request"));
-        spans.add(new Span(mock(ElasticApmTracer.class))
+        spans.add(new Span(MockTracer.create())
             .start(TraceContext.fromParent(), t, -1, false)
             .withName("GET /api/types")
             .withType("request"));
-        spans.add(new Span(mock(ElasticApmTracer.class))
+        spans.add(new Span(MockTracer.create())
             .start(TraceContext.fromParent(), t, -1, false)
             .withName("GET /api/types")
             .withType("request"));
 
-        span = new Span(mock(ElasticApmTracer.class))
+        span = new Span(MockTracer.create())
             .start(TraceContext.fromParent(), t, -1, false)
             .appendToName("GET ")
             .appendToName("test.elastic.co")