+++ apm-agent-core/src/test/java/co/elastic/apm/agent/metrics/MetricRegistryTest.java
@@ -11,9 +11,9 @@
  * the Apache License, Version 2.0 (the "License"); you may
  * not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -30,8 +30,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
+import java.util.stream.IntStream;
 
-import static java.util.Collections.emptyMap;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -53,9 +53,27 @@ class MetricRegistryTest {
         final DoubleSupplier problematicMetric = () -> {
             throw new RuntimeException("Huston, we have a problem");
         };
-        metricRegistry.addUnlessNegative("jvm.gc.count", emptyMap(), problematicMetric);
-        metricRegistry.addUnlessNan("jvm.gc.count", emptyMap(), problematicMetric);
-        metricRegistry.add("jvm.gc.count", emptyMap(), problematicMetric);
-        assertThat(metricRegistry.getMetricSets()).isEmpty();
+        metricRegistry.addUnlessNegative("jvm.gc.count", Labels.EMPTY, problematicMetric);
+        metricRegistry.addUnlessNan("jvm.gc.count", Labels.EMPTY, problematicMetric);
+        metricRegistry.add("jvm.gc.count", Labels.EMPTY, problematicMetric);
+        metricRegistry.report(metricSets -> assertThat(metricSets).isEmpty());
+    }
+
+    @Test
+    void testReportGaugeTwice() {
+        metricRegistry.add("foo", Labels.EMPTY, () -> 42);
+        metricRegistry.report(metricSets -> assertThat(metricSets.get(Labels.EMPTY).getGauge("foo").get()).isEqualTo(42));
+        // the active and inactive metricSets are now switched, verify that the previous inactive metricSets also contain the same gauges
+        metricRegistry.report(metricSets -> assertThat(metricSets.get(Labels.EMPTY).getGauge("foo").get()).isEqualTo(42));
+    }
+
+    @Test
+    void testLimitTimers() {
+        IntStream.range(1, 505).forEach(i -> metricRegistry.updateTimer("timer" + i, Labels.Mutable.of("foo", Integer.toString(i)), 1));
+        IntStream.range(1, 505).forEach(i -> metricRegistry.updateTimer("timer" + i, Labels.Mutable.of("bar", Integer.toString(i)), 1));
+
+        metricRegistry.report(metricSets -> assertThat(metricSets).hasSize(1000));
+        // the active and inactive metricSets are now switched, also check the size of the previously inactive metricSets
+        metricRegistry.report(metricSets -> assertThat(metricSets).hasSize(1000));
     }
 }